BaseInterwikiResolver::loadConfigFromApi already returns array
The getSameProjectConfigByLang was wrapping the config call in an
array to ttach the prefix, but loadConfigFromAPI already did this.
Having it wrapped twice caused BadMethodCall exceptions as we
tried to treat the array as if it was a SearchConfig object.

Additionally in the main language detection codepath there
was an incorrect check for empty( $interwikiConfig ) but that
variable doesn't exist anymore. Update to check the appropriate
new variable, $iwPrefixAndConfig.

Bug: T177535
Change-Id: I05739753c07d89942685f0e55d372a870bc9926c